Cargo.toml fifo_data.rs trigger.rs var.rs errors.rs file_upwards.rs string_upwards.rs

+++ src/client/fifo_data.rs
@@ -8,7 +8,7 @@ use serde::Deserialize;
 pub struct FifoTrigger {
   pub command: String,
   pub file: Option<String>,
-  pub line: Option<u32>,
+  pub line: Option<usize>,
   pub run: Option<String>,
 }
